Split out DefaultConfiguration
